server/requirements.txt
# requirements.txt
Django==4.2
djangorestframework==3.14.0
daphne==4.1.2
channels==4.0.0
channels-redis==4.0.0
django-cors-headers==4.7.0
